refactor: Clean up evaluation call to rule engine
