--- src/commonMain/kotlin/org/hisp/dhis/rules/DefaultRuleEngineAPI.kt
@@ -1,42 +1,0 @@
-package org.hisp.dhis.rules
-
-import org.hisp.dhis.lib.expression.Expression
-import org.hisp.dhis.lib.expression.spi.IllegalExpressionException
-import org.hisp.dhis.lib.expression.spi.ParseException
-import org.hisp.dhis.lib.expression.spi.ValueType
-import org.hisp.dhis.rules.models.RuleEngineValidationException
-import org.hisp.dhis.rules.models.RuleValidationResult
-
-class DefaultRuleEngineAPI: RuleEngineAPI {
-    override fun validate(expression: String, dataItemStore: Map<String, DataItem>): RuleValidationResult {
-        // Rule condition expression should be evaluated against Boolean
-        return getExpressionDescription(expression, Expression.Mode.RULE_ENGINE_CONDITION, dataItemStore)
-    }
-
-    override fun validateDataFieldExpression(expression: String, dataItemStore: Map<String, DataItem>): RuleValidationResult {
-        // Rule action data field should be evaluated against all i.e Boolean, String, Date and Numerical value
-        return getExpressionDescription(expression, Expression.Mode.RULE_ENGINE_ACTION, dataItemStore)
-    }
-
-    private fun getExpressionDescription(expression: String, mode: Expression.Mode, dataItemStore: Map<String, DataItem>): RuleValidationResult {
-        return try {
-            val validationMap: Map<String, ValueType> = dataItemStore.mapValues { e -> e.value.valueType.toValueType() }
-            Expression(expression, mode, false).validate(validationMap)
-            val displayNames: Map<String, String> = dataItemStore.mapValues { e -> e.value.displayName }
-            val description = Expression(expression, mode, false).describe(displayNames)
-            RuleValidationResult(valid = true, description = description)
-        } catch (ex: IllegalExpressionException) {
-            RuleValidationResult(
-                valid = false,
-                exception = RuleEngineValidationException(ex),
-                errorMessage = ex.message
-            )
-        } catch (ex: ParseException) {
-            RuleValidationResult(
-                valid = false,
-                exception = RuleEngineValidationException(ex),
-                errorMessage = ex.message
-            )
-        }
-    }
-}
