+++ src/commonMain/kotlin/org/hisp/dhis/rules/RuleEngine.kt
@@ -6,9 +6,9 @@ import org.hisp.dhis.lib.expression.spi.ParseException
 import org.hisp.dhis.lib.expression.spi.ValueType
 import org.hisp.dhis.rules.models.*
 
-class RuleEngine {
+class RuleEngine: RuleEngineAPI {
 
-    fun evaluate(ruleEvent: RuleEvent, executionContext: RuleEngineContext): List<RuleEffect> {
+    override fun evaluate(ruleEvent: RuleEvent, executionContext: RuleEngineContext): List<RuleEffect> {
         val valueMap = RuleVariableValueMapBuilder.target(ruleEvent)
             .ruleVariables(executionContext.ruleVariables)
             .ruleEnrollment(executionContext.enrollment)
@@ -19,7 +19,7 @@ class RuleEngine {
         return RuleEngineExecution(ruleEvent, null, executionContext.rules, valueMap, executionContext.supplementaryData).execute()
     }
 
-    fun evaluate(ruleEnrollment: RuleEnrollment, executionContext: RuleEngineContext): List<RuleEffect> {
+    override fun evaluate(ruleEnrollment: RuleEnrollment, executionContext: RuleEngineContext): List<RuleEffect> {
         val valueMap = RuleVariableValueMapBuilder.target(ruleEnrollment)
                 .ruleVariables(executionContext.ruleVariables)
                 .triggerEnvironment(TriggerEnvironment.SERVER)
@@ -29,7 +29,7 @@ class RuleEngine {
         return RuleEngineExecution(null, ruleEnrollment, executionContext.rules, valueMap, executionContext.supplementaryData).execute()
     }
 
-    fun evaluate(executionContext: RuleEngineContext): List<RuleEffects> {
+    override fun evaluate(executionContext: RuleEngineContext): List<RuleEffects> {
         val valueMap = RuleVariableValueMapBuilder.target()
                 .ruleVariables(executionContext.ruleVariables)
                 .ruleEnrollment(executionContext.enrollment)
@@ -39,5 +39,37 @@ class RuleEngine {
                 .multipleBuild()
         return RuleEngineMultipleExecution(executionContext.rules, valueMap,
                 executionContext.supplementaryData).execute()
+    }
+
+    override fun validate(expression: String, dataItemStore: Map<String, DataItem>): RuleValidationResult {
+        // Rule condition expression should be evaluated against Boolean
+        return getExpressionDescription(expression, Expression.Mode.RULE_ENGINE_CONDITION, dataItemStore)
+    }
+
+    override fun validateDataFieldExpression(expression: String, dataItemStore: Map<String, DataItem>): RuleValidationResult {
+        // Rule action data field should be evaluated against all i.e Boolean, String, Date and Numerical value
+        return getExpressionDescription(expression, Expression.Mode.RULE_ENGINE_ACTION, dataItemStore)
+    }
+
+    private fun getExpressionDescription(expression: String, mode: Expression.Mode, dataItemStore: Map<String, DataItem>): RuleValidationResult {
+        return try {
+            val validationMap: Map<String, ValueType> = dataItemStore.mapValues { e -> e.value.valueType.toValueType() }
+            Expression(expression, mode, false).validate(validationMap)
+            val displayNames: Map<String, String> = dataItemStore.mapValues { e -> e.value.displayName }
+            val description = Expression(expression, mode, false).describe(displayNames)
+            RuleValidationResult(valid = true, description = description)
+        } catch (ex: IllegalExpressionException) {
+            RuleValidationResult(
+                valid = false,
+                exception = RuleEngineValidationException(ex),
+                errorMessage = ex.message
+            )
+        } catch (ex: ParseException) {
+            RuleValidationResult(
+                valid = false,
+                exception = RuleEngineValidationException(ex),
+                errorMessage = ex.message
+            )
+        }
     }
 }
